requirements.txt
azure-core==1.30.1
azure-storage-blob==12.19.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
elastic-transport==8.12.0
elasticsearch==8.12.1
idna==3.6
isodate==0.6.1
lxml==5.1.0
numpy==1.26.4
pandas==2.2.1
pycparser==2.21
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
